+++ common-api-operation/src/main/java/com/gb/tech/apioperation/ApiOperationApplication.java
@@ -1,0 +1,13 @@
+package com.gb.tech.apioperation;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ApiOperationApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ApiOperationApplication.class, args);
+	}
+
+}
