+++ common-api-operation/src/main/java/com/gb/tech/apioperation/exeptions/AppError.java
@@ -1,0 +1,15 @@
+package com.gb.tech.apioperation.exeptions;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+
+public class AppError {
+    private String statusCode;
+    private String message;
+}
