+++ common-api-operation/src/main/java/com/gb/tech/apioperation/exeptions/InvalidParamsException.java
@@ -1,0 +1,7 @@
+package com.gb.tech.apioperation.exeptions;
+
+public class InvalidParamsException extends RuntimeException {
+    public InvalidParamsException(String message) {
+        super(message);
+    }
+}
