+++ common-api-operation/src/main/java/com/gb/tech/apioperation/exeptions/ServiceErrors.java
@@ -1,0 +1,9 @@
+package com.gb.tech.apioperation.exeptions;
+
+public enum ServiceErrors {
+    NOT_FOUND,
+    VALIDATION_ERRORS,
+    INVALID_PARAMS,
+    INCORRECT_TOKEN,
+    SERVICE_UNAVAILABLE
+}
