+++ common-api-operation/src/main/java/com/gb/tech/apioperation/repository/FinancialOperationRepository.java
@@ -1,0 +1,12 @@
+package com.gb.tech.apioperation.repository;
+
+import com.gb.tech.apioperation.entity.FinancialOperation;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface FinancialOperationRepository extends JpaRepository<FinancialOperation, Long> {
+
+    List<FinancialOperation> findAllByIsSpending(Boolean isSpending);
+
+}
