+++ common-api-operation/src/main/java/com/gb/tech/apioperation/service/FinancialOperationServiceImp.java
@@ -1,0 +1,54 @@
+package com.gb.tech.apioperation.service;
+
+import com.gb.tech.apioperation.exeptions.InvalidParamsException;
+import com.gb.tech.apioperation.repository.FinancialOperationRepository;
+import com.gb.tech.apioperation.entity.FinancialOperation;
+import com.gb.tech.apioperation.dto.FinancialOperationDto;
+import com.gb.tech.apioperation.mapper.FinancialOperationMapper;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+
+import javax.transaction.Transactional;
+import java.util.List;
+
+
+@Service
+@AllArgsConstructor
+public class FinancialOperationServiceImp implements FinancialOperationService{
+
+    private final FinancialOperationMapper mapper = FinancialOperationMapper.MAPPER;
+
+    private final FinancialOperationRepository financialOperationRepository;
+
+    @Override
+    public FinancialOperationDto getById(Long id){
+        FinancialOperation financialOperation = financialOperationRepository.findById(id).orElse(null);
+        // с обработкой ошибок определимся позже.
+        if (financialOperation == null) {
+            throw new InvalidParamsException("financialOperation:" + null);
+        }
+        return mapper.fromFinancialOperation(financialOperation);
+    }
+
+    @Override
+    public List<FinancialOperationDto> getAll(Boolean isSpending){
+        return mapper.fromFinancialOperationList(financialOperationRepository.findAllByIsSpending(isSpending));
+    }
+
+    @Override
+    @Transactional
+    public void save(FinancialOperationDto financialOperationDto) {
+
+        financialOperationRepository.save(mapper.toFinancialOperation(financialOperationDto));
+
+    }
+
+    @Override
+    public void delete(Long id){
+
+        financialOperationRepository.deleteById(id);
+    }
+
+
+}
