+++ common-api-operation/src/main/resources/application.yaml
@@ -1,0 +1,23 @@
+spring:
+  jpa:
+    hibernate:
+      ddl-auto: validate
+    show-sql: true
+  application:
+    name:
+      COMMON-API-OPERATION
+#  datasource:
+#    url: jdbc:postgresql://localhost:5432/finasist
+#    username: geek
+#    password: geek
+  datasource:
+    url: "jdbc:postgresql://postgres/finasist"
+#           добавить базу
+
+#  eureka:
+#    client:
+#      serviceUrl:
+#        defaultZone:
+#          http://common-eureka-spring:8761/eureka/
+#    instance:
+#      preferIpAddress: true
