+++ common-api-operation/src/test/java/com/gb/tech/apioperation/FinancialAssistantApplicationTests.java
@@ -1,0 +1,13 @@
+package com.gb.tech.apioperation;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class FinancialAssistantApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
