--- financial-assistant/src/main/java/com/gb/tech/financialassistant/FinancialAssistantApplication.java
@@ -1,13 +1,0 @@
-package com.gb.tech.financialassistant;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class FinancialAssistantApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(FinancialAssistantApplication.class, args);
-	}
-
-}
