--- financial-assistant/src/main/java/com/gb/tech/financialassistant/exeptions/AppError.java
@@ -1,15 +1,0 @@
-package com.gb.tech.financialassistant.exeptions;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-
-public class AppError {
-    private String statusCode;
-    private String message;
-}
