--- financial-assistant/src/main/java/com/gb/tech/financialassistant/exeptions/InvalidParamsException.java
@@ -1,7 +1,0 @@
-package com.gb.tech.financialassistant.exeptions;
-
-public class InvalidParamsException extends RuntimeException {
-    public InvalidParamsException(String message) {
-        super(message);
-    }
-}
