--- financial-assistant/src/main/java/com/gb/tech/financialassistant/exeptions/ServiceErrors.java
@@ -1,9 +1,0 @@
-package com.gb.tech.financialassistant.exeptions;
-
-public enum ServiceErrors {
-    NOT_FOUND,
-    VALIDATION_ERRORS,
-    INVALID_PARAMS,
-    INCORRECT_TOKEN,
-    SERVICE_UNAVAILABLE
-}
