--- financial-assistant/src/main/java/com/gb/tech/financialassistant/repository/FinancialOperationRepository.java
@@ -1,12 +1,0 @@
-package com.gb.tech.financialassistant.repository;
-
-import com.gb.tech.financialassistant.entity.FinancialOperation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface FinancialOperationRepository extends JpaRepository<FinancialOperation, Long> {
-
-    List<FinancialOperation> findAllByIsSpending(Boolean isSpending);
-
-}
