--- financial-assistant/src/main/resources/application.yaml
@@ -1,11 +1,0 @@
-spring:
-  jpa:
-    hibernate:
-      ddl-auto: validate
-    show-sql: true
-  datasource:
-    url: jdbc:postgresql://localhost:5432/finasist
-    username: geek
-    password: geek
-#  flyway:
-#    baseline-on-migrate: true
