--- financial-assistant/src/test/java/com/gb/tech/financialassistant/FinancialAssistantApplicationTests.java
@@ -1,13 +1,0 @@
-package com.gb.tech.financialassistant;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FinancialAssistantApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
