+++ lib-exception/pom.xml
@@ -4,7 +4,7 @@
 
   <groupId>com.gb.financial.assistant</groupId>
   <artifactId>lib-exception</artifactId>
-  <version>1.0.0</version>
+  <version>1.0.1</version>
   <packaging>jar</packaging>
 
   <name>lib-exception</name>
